Switched to send request object as JSON to proxy
