+++ connectors/http-json/src/main/java/io/camunda/connector/http/HttpJsonFunction.java
@@ -16,37 +16,17 @@
  */
 package io.camunda.connector.http;
 
-import static org.apache.http.entity.ContentType.APPLICATION_JSON;
-
-import com.google.api.client.http.GenericUrl;
-import com.google.api.client.http.HttpContent;
-import com.google.api.client.http.HttpHeaders;
-import com.google.api.client.http.HttpRequest;
 import com.google.api.client.http.HttpRequestFactory;
-import com.google.api.client.http.HttpResponse;
-import com.google.api.client.http.HttpResponseException;
-import com.google.api.client.http.json.JsonHttpContent;
 import com.google.api.client.json.gson.GsonFactory;
 import com.google.gson.Gson;
 import io.camunda.connector.api.annotation.OutboundConnector;
-import io.camunda.connector.api.error.ConnectorException;
 import io.camunda.connector.api.outbound.OutboundConnectorContext;
 import io.camunda.connector.api.outbound.OutboundConnectorFunction;
 import io.camunda.connector.http.components.GsonComponentSupplier;
 import io.camunda.connector.http.components.HttpTransportComponentSupplier;
 import io.camunda.connector.http.model.HttpJsonRequest;
-import io.camunda.connector.http.model.HttpJsonResult;
-import io.camunda.connector.impl.ConnectorInputException;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.TimeUnit;
-import javax.validation.ValidationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -61,15 +41,14 @@ import org.slf4j.LoggerFactory;
       "connectionTimeoutInSeconds",
       "body"
     },
-    type = "io.camunda:http-json:1")
+    type = "io.camunda:http-json:test")
 public class HttpJsonFunction implements OutboundConnectorFunction {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(HttpJsonFunction.class);
   protected static final String REQUEST_HEADER_TARGET_URL = "X-Camunda-Target-URL";
 
   private final Gson gson;
-  private final GsonFactory gsonFactory;
-  private final HttpRequestFactory requestFactory;
+  private final HttpInvocationHelper http;
 
   private Optional<String> proxyFunctionToUse = Optional.empty();
 
@@ -83,10 +62,9 @@ public class HttpJsonFunction implements OutboundConnectorFunction {
   public HttpJsonFunction(
       final Gson gson, final HttpRequestFactory requestFactory, final GsonFactory gsonFactory) {
     this.gson = gson;
-    this.requestFactory = requestFactory;
-    this.gsonFactory = gsonFactory;
+    this.http = new HttpInvocationHelper(gson, requestFactory, gsonFactory);
 
-    proxyFunctionToUse = Optional.of("http://localhost:8080/function-connector-http-proxy-function-test1");
+    //proxyFunctionToUse = Optional.of("http://localhost:8080/function-connector-http-proxy-function-test1");
   }
 
   @Override
@@ -97,110 +75,10 @@ public class HttpJsonFunction implements OutboundConnectorFunction {
     context.validate(request);
     context.replaceSecrets(request);
 
-    return handleRequest(request);
-  }
-
-  protected HttpJsonResult handleRequest(final HttpJsonRequest request) throws IOException {
-    final HttpRequest externalRequest = createRequest(request);
-    final HttpResponse externalResponse = sendRequest(externalRequest);
-    return toHttpJsonResponse(externalResponse);
-  }
-
-  protected HttpResponse sendRequest(final HttpRequest request) throws IOException {
-    try {
-      return request.execute();
-    } catch (HttpResponseException hrex) {
-      throw new ConnectorException(String.valueOf(hrex.getStatusCode()), hrex.getMessage());
-    }
-  }
-
-  private HttpRequest createRequest(final HttpJsonRequest request) throws IOException {
-    final String url = proxyFunctionToUse.orElse(request.getUrl());
-    // TODO: Decide if we want to keep it now that we can use a proxy
-    if (url.contains("computeMetadata")) {
-      throw new ConnectorInputException(new ValidationException("The provided URL is not allowed"));
-    }
-
-    final var content = createContent(request);
-    final var method = request.getMethod().toUpperCase();
-
-    final GenericUrl genericUrl = new GenericUrl(url);
-
-    if (request.hasQueryParameters()) {
-      genericUrl.putAll(request.getQueryParameters());
-    }
-
-    final var httpRequest = requestFactory.buildRequest(method, genericUrl, content);
-    httpRequest.setFollowRedirects(false);
-
-    if (request.getConnectionTimeoutInSeconds() != null) {
-      long connectionTimeout =
-          TimeUnit.SECONDS.toMillis(Long.parseLong(request.getConnectionTimeoutInSeconds()));
-      int intConnectionTimeout = Math.toIntExact(connectionTimeout);
-      httpRequest.setConnectTimeout(intConnectionTimeout);
-      httpRequest.setReadTimeout(intConnectionTimeout);
-      httpRequest.setWriteTimeout(intConnectionTimeout);
-    }
-
-    final var headers = createHeaders(request);
     if (proxyFunctionToUse.isPresent()) {
-      headers.set(REQUEST_HEADER_TARGET_URL, request.getUrl());
-    }
-    httpRequest.setHeaders(headers);
-
-    return httpRequest;
-  }
-
-  private HttpContent createContent(final HttpJsonRequest request) {
-    if (request.hasBody()) {
-      return new JsonHttpContent(gsonFactory, request.getBody());
+      return http.executeRequestViaProxy(proxyFunctionToUse.get(), request);
     } else {
-      return null;
+      return http.executeRequestDirectly(request);
     }
-  }
-
-  private HttpHeaders createHeaders(final HttpJsonRequest request) {
-    final HttpHeaders httpHeaders = new HttpHeaders();
-
-    if (request.hasBody()) {
-      httpHeaders.setContentType(APPLICATION_JSON.getMimeType());
-    }
-
-    if (request.hasAuthentication()) {
-      request.getAuthentication().setHeaders(httpHeaders);
-    }
-
-    if (request.hasHeaders()) {
-      httpHeaders.putAll(request.getHeaders());
-    }
-
-    return httpHeaders;
-  }
-
-  private HttpJsonResult toHttpJsonResponse(final HttpResponse externalResponse) {
-    final HttpJsonResult httpJsonResult = new HttpJsonResult();
-    httpJsonResult.setStatus(externalResponse.getStatusCode());
-    final Map<String, Object> headers = new HashMap<>();
-    externalResponse
-        .getHeaders()
-        .forEach(
-            (k, v) -> {
-              if (v instanceof List && ((List<?>) v).size() == 1) {
-                headers.put(k, ((List<?>) v).get(0));
-              } else {
-                headers.put(k, v);
-              }
-            });
-    httpJsonResult.setHeaders(headers);
-    try (InputStream content = externalResponse.getContent();
-        Reader reader = new InputStreamReader(content)) {
-      final Object body = gson.fromJson(reader, Object.class);
-      if (body != null) {
-        httpJsonResult.setBody(body);
-      }
-    } catch (final Exception e) {
-      LOGGER.error("Failed to parse external response: {}", externalResponse, e);
-    }
-    return httpJsonResult;
   }
 }
