+++ connectors/http-json/src/test/java/io/camunda/connector/http/LocalConnectorRuntime.java
@@ -23,8 +23,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 //@SpringBootApplication
 public class LocalConnectorRuntime {
 
-    public static void main(String[] args) {
-        SpringApplication.run(ConnectorRuntimeApplication.class, args);
-    }
-
+  public static void main(String[] args) {
+    SpringApplication.run(ConnectorRuntimeApplication.class, args);
+  }
 }
